+++ src/typescript/actor/wvActorSheet.ts
@@ -15,7 +15,7 @@ export default class WvActorSheet extends ActorSheet<
 > {
   static override get defaultOptions(): ActorSheet.Options {
     return foundry.utils.mergeObject(super.defaultOptions, {
-      classes: [CONSTANTS.systemId, "actor-sheet"],
+      classes: [CONSTANTS.systemId, "entity-sheet", "actor-sheet"],
       tabs: [
         { navSelector: ".tabs", contentSelector: ".content", initial: "stats" }
       ],