+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - flavor text for character sheet rolls ([#1](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/1))
 - translatable migration notifications
 - clickable labels to focus the corresponding target inputs
+- roll with modifier by Shift-Clicking the roll buttons  ([#2](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/2))
+- roll as GM whisper by Ctrl-Clicking the roll buttons
 
 ### Fixed