+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- not being able to save the character sheet due to the new validation
+
 ## [0.3.2] - 2021-08-07
 
 ### Added