+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.3] - 2021-08-08
+
+### Added
+
+- configurable minimum bounds for SPECIALs and Skills
+
 ### Fixed
 
 - not being able to save the character sheet due to the new validation
@@ -106,7 +112,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - basic character Actor implementation with editable SPECIALs
 
-[Unreleased]: https://github.com/Wasteland-Ventures-Group/WV-VTT-module/compare/v0.3.2...HEAD
+[Unreleased]: https://github.com/Wasteland-Ventures-Group/WV-VTT-module/compare/v0.3.3...HEAD
+[0.3.3]:      https://github.com/Wasteland-Ventures-Group/WV-VTT-module/compare/v0.3.2...v0.3.3
 [0.3.2]:      https://github.com/Wasteland-Ventures-Group/WV-VTT-module/compare/v0.3.1...v0.3.2
 [0.3.1]:      https://github.com/Wasteland-Ventures-Group/WV-VTT-module/compare/v0.3.0...v0.3.1
 [0.3.0]:      https://github.com/Wasteland-Ventures-Group/WV-VTT-module/compare/v0.2.0...v0.3.0